pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "horizon-infra-connectors"
version = "0.0.0.0.dev1"
description = "Async helpers that wrap Argo CD, Git providers, and Vault APIs."
readme = "README.md"
authors = [
    { name = "Yehonatan Magen" },
    { name = "Alon Elimelech" }
]
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.95",
    "httpx>=0.24",
    "loguru>=0.7",
    "pyyaml>=6.0.3",
    "pydantic>=2.0",
    "pydantic-settings>=2.0"
]

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools]
include-package-data = true
